chore: fix test and lint issues
