+++ packages/env-capabilities/bundlesize.json
@@ -2,32 +2,32 @@
   "files": [
     {
       "path": "dist/es/index.debug.js",
-      "maxSize": 201,
+      "maxSize": 89,
       "compression": "none"
     },
     {
       "path": "dist/cjs/index.debug.js",
-      "maxSize": 202,
+      "maxSize": 89,
       "compression": "none"
     },
     {
       "path": "dist/iife/index.debug.js",
-      "maxSize": 207,
-      "compression": "none"
-    },
-    {
-      "path": "dist/cjs/index.min.js",
-      "maxSize": 30,
+      "maxSize": 91,
       "compression": "none"
     },
     {
       "path": "dist/es/index.min.js",
-      "maxSize": 30,
+      "maxSize": 9,
+      "compression": "none"
+    },
+    {
+      "path": "dist/cjs/index.min.js",
+      "maxSize": 9,
       "compression": "none"
     },
     {
       "path": "dist/iife/index.min.js",
-      "maxSize": 30,
+      "maxSize": 9,
       "compression": "none"
     }
   ]
