+++ packages/hls-parser/demo/src/main.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-import { FullPlaylistParser, ProgressiveParser } from '../../src';
+import { FullPlaylistParser, ChunkPlaylistParser } from '../../src';
 
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
@@ -7,7 +7,7 @@ import App from './App';
 import './index.css';
 
 console.dir(FullPlaylistParser);
-console.dir(ProgressiveParser);
+console.dir(ChunkPlaylistParser);
 
 const root = document.getElementById('root')!;
