+++ packages/playback/src/lib/consts/playback-state.ts
@@ -2,6 +2,7 @@
 // since they can be used as values they should not be in the types folder
 export enum PlaybackState {
   Idle = 'Idle',
+  Loading = 'Loading',
   Playing = 'Playing',
   Paused = 'Paused',
   Buffering = 'Buffering',
