+++ packages/playback/src/lib/player/base/base-player.ts
@@ -2,8 +2,7 @@
 import { LoggerLevel } from '../../consts/logger-level';
 import { PlayerEventType } from '../../consts/events';
 // types
-import type { IPlayerSource } from '../../types/source.declarations';
-import type { IInterceptorsStorage } from '../../types/interceptors.declarations';
+import type { ILoadLocalSource, ILoadRemoteSource, IPlayerSource } from '../../types/source.declarations';
 import type { IInterceptorsStorage, Interceptor } from '../../types/interceptors.declarations';
 import type { ILogger } from '../../types/logger.declarations';
 import type { PlayerConfiguration } from '../../types/configuration.declarations';
@@ -12,13 +11,29 @@ import type { IStore } from '../../types/store.declarations';
 import type { EventListener, IEventEmitter } from '../../types/event-emitter.declarations';
 import type { EventTypeToEventMap } from '../../types/mappers/event-type-to-event-map.declarations';
 // events
-import { ConfigurationChangedEvent, LoggerLevelChangedEvent, VolumeChangedEvent } from '../../events/player-events';
+import {
+  ConfigurationChangedEvent,
+  LoggerLevelChangedEvent,
+  MutedStatusChangedEvent,
+  PlayerErrorEvent,
+  VolumeChangedEvent,
+} from '../../events/player-events';
 // models
+import { PlayerSource } from '../../models/player-source';
 // errors
+import { NoSupportedPipelineError } from '../../errors/pipeline-errors';
 // pipelines
 import { PipelineLoaderFactoryStorage } from './pipeline-loader-factory-storage';
 import type { InterceptorType } from '../../consts/interceptor-type';
 import type { InterceptorTypeToInterceptorPayloadMap } from '../../types/mappers/interceptor-type-to-interceptor-map.declarations';
+import { PlaybackState } from '../../consts/playback-state';
+import { PlayerTimeRange } from '../../models/player-time-range';
+import type { IPlayerTimeRange } from '../../types/player-time-range.declarations';
+import type { IPlaybackStats } from '../../types/playback-stats.declarations';
+import type { IPlayerAudioTrack } from '../../types/audio-track.declarations';
+import type { IPlayerThumbnailTrack, IRemoteVttThumbnailTrackOptions } from '../../types/thumbnail-track.declarations';
+import type { IQualityLevel } from '../../types/quality-level.declarations';
+import type { VersionInfo } from '../../types/version-info.declarations';
 
 declare const __COMMIT_HASH: string;
 declare const __VERSION: string;
@@ -47,22 +62,18 @@ export abstract class BasePlayer {
   /**
    * static player's event type enum getter
    */
-  public static readonly Event = PlayerEventType;
+  public static readonly EventType = PlayerEventType;
 
   /**
-   * static player's version getter
+   * static player's version info getter
    */
-  public static readonly version: string = __VERSION;
-
-  /**
-   * static player's version hash getter
-   */
-  public static readonly versionHash: string = __COMMIT_HASH;
-
-  /**
-   * static is experimental build flag getter
-   */
-  public static readonly isExperimentalBuild: boolean = __EXPERIMENTAL;
+  public static get versionIfo(): VersionInfo {
+    return {
+      version: __VERSION,
+      versionHash: __COMMIT_HASH,
+      isExperimental: __EXPERIMENTAL,
+    };
+  }
 
   /**
    * static pipeline loader factory storage getter
@@ -72,6 +83,8 @@ export abstract class BasePlayer {
   /**
    * MARK: PROTECTED INSTANCE MEMBERS
    */
+
+  protected currentPlaybackState_: PlaybackState = PlaybackState.Idle;
 
   /**
    * attached video element
@@ -107,6 +120,10 @@ export abstract class BasePlayer {
    */
   private readonly configurationManager_: IStore<PlayerConfiguration>;
 
+  /**
+   * You can pass your own implementations via dependencies.
+   * @param dependencies - player dependencies
+   */
   protected constructor(dependencies: PlayerDependencies) {
     this.logger_ = dependencies.logger;
     this.eventEmitter_ = dependencies.eventEmitter;
@@ -158,7 +175,7 @@ export abstract class BasePlayer {
   }
 
   /**
-   * MARK: Debug API
+   * MARK: DEBUG API
    */
 
   /**
@@ -178,7 +195,7 @@ export abstract class BasePlayer {
   }
 
   /**
-   * MARK: Configuration API
+   * MARK: CONFIGURATION API
    */
 
   /**
@@ -208,7 +225,7 @@ export abstract class BasePlayer {
   }
 
   /**
-   * MARK: Events API
+   * MARK: EVENTS API
    */
 
   /**
@@ -259,7 +276,7 @@ export abstract class BasePlayer {
   }
 
   /**
-   * MARK: Life Cycle API
+   * MARK: LIFE CYCLE API
    */
 
   /**
@@ -295,6 +312,7 @@ export abstract class BasePlayer {
 
     this.stop('detach');
 
+    // TODO: update list of all possible events + add EME events here as well
     this.activeVideoElement_.removeEventListener('volumechange', this.handleVolumeChange_);
 
     this.activeVideoElement_ = null;
@@ -309,6 +327,7 @@ export abstract class BasePlayer {
 
     this.activeSource_?.dispose();
     this.activeSource_ = null;
+    this.transitionPlaybackState_(PlaybackState.Idle);
 
     // this.activePipelineLoader_?.abort();
     // this.activePipelineLoader_ = null;
@@ -323,6 +342,78 @@ export abstract class BasePlayer {
     this.detach();
     this.eventEmitter_.removeAllEventListeners();
     this.interceptorsStorage_.removeAllInterceptors();
+  }
+
+  /**
+   * load source
+   * @param source - source to load
+   */
+  public load(source: ILoadRemoteSource | ILoadLocalSource): void {
+    if (this.activeVideoElement_ === null) {
+      this.logger_.warn(
+        'You are trying to load a source, while no video element attached to the player. Please call "attach" with a target video element first.'
+      );
+
+      return;
+    }
+
+    if (this.activeSource_ !== null) {
+      this.stop('load');
+    }
+
+    this.transitionPlaybackState_(PlaybackState.Loading);
+    this.activeSource_ = new PlayerSource(source);
+
+    this.logger_.debug('received a load request: ', this.activeSource_);
+
+    // TODO: implement pipeline loader flow
+    // if (this.hasPipelineFactoryConfiguration(this.activeSource_.mimeType.toLowerCase())) {
+    //   const { loader, vod, live } = this.getPipelineFactoryConfiguration(this.activeSource_.mimeType.toLowerCase())!;
+    //   this.activePipelineLoader_ = loader.create({
+    //     logger: this.logger_,
+    //     videoElement: this.activeVideoElement_,
+    //     networkManager: this.networkManager_,
+    //     source: this.activeSource_,
+    //     vodFactory: vod,
+    //     liveFactory: live,
+    //   });
+    //
+    //   this.activePipelineLoader_.load().then(
+    //     (pipeline) => {
+    //       this.activePipeline_ = pipeline;
+    //     },
+    //     (error) => {
+    //       this.eventEmitter_.emitEvent(
+    //         new PlayerErrorEvent(new PipelineLoaderFailedToDeterminePipelineError(true, error))
+    //       );
+    //     }
+    //   );
+    //
+    //   return;
+    // }
+
+    this.logger_.debug(
+      `No registered pipeline's configuration found for the provided mime type (${this.activeSource_.mimeType}).`
+    );
+
+    if (this.activeVideoElement_.canPlayType(this.activeSource_.mimeType)) {
+      this.logger_.debug('Native Pipeline can play the provided mime type. Fallback to the Native Pipeline');
+      // TODO: implement NativePipeline flow
+      // this.activePipeline_ = NativePipeline.create({
+      //   logger: this.logger_,
+      //   videoElement: this.activeVideoElement_,
+      //   networkManager: this.networkManager_,
+      //   source: this.activeSource_,
+      // });
+      //
+      // this.activePipeline_.start();
+      return;
+    } else {
+      this.logger_.debug('Native Pipeline can not play the provided mime type.');
+    }
+
+    this.logger_.warn('No supported pipelines found for ', source.mimeType);
+    this.eventEmitter_.emitEvent(new PlayerErrorEvent(new NoSupportedPipelineError(true)));
   }
 
   /**
@@ -341,4 +432,310 @@ export abstract class BasePlayer {
 
     this.eventEmitter_.emitEvent(new VolumeChangedEvent(target.volume));
   };
+
+  private transitionPlaybackState_(to: PlaybackState): void {
+    if (this.currentPlaybackState_ !== to) {
+      this.currentPlaybackState_ = to;
+      // TODO: emit event
+    }
+  }
+
+  /**
+   * MARK: Playback API
+   */
+
+  /**
+   * Resumes active playback session (if any)
+   */
+  public play(): void {
+    this.activeVideoElement_?.play();
+    // TODO: pass command to loader/pipeline
+  }
+
+  /**
+   * Pauses active playback session (if any)
+   */
+  public pause(): void {
+    this.activeVideoElement_?.pause();
+    // TODO: pass command to loader/pipeline
+  }
+
+  /**
+   * seek to the provided seek target
+   * should return false if seek target is out of available seekable ranges
+   * @param seekTarget - seek target
+   */
+  public seek(seekTarget: number): boolean {
+    if (!this.activeVideoElement_) {
+      return false;
+    }
+
+    const seekableRanges = this.getSeekableRanges();
+    const isValidSeekTarget = seekableRanges.some((timeRange) => timeRange.isInRangeInclusive(seekTarget));
+
+    if (!isValidSeekTarget) {
+      this.logger_.warn(
+        `Provided seek target (${seekTarget}) is out of available seekable time ranges: `,
+        seekableRanges.map((range) => range.toString()).join('\n')
+      );
+      return false;
+    }
+
+    this.activeVideoElement_.currentTime = seekTarget;
+    // TODO: pass command to loader/pipeline
+    return true;
+  }
+
+  /**
+   * current time getter
+   */
+  public getCurrentTime(): number {
+    return this.activeVideoElement_?.currentTime ?? 0;
+  }
+
+  /**
+   * playback rate getter
+   */
+  public getPlaybackRate(): number {
+    return this.activeVideoElement_?.playbackRate ?? 1;
+  }
+
+  /**
+   * playback rate setter
+   * @param rate - playback rate
+   */
+  public setPlaybackRate(rate: number): void {
+    if (this.activeVideoElement_) {
+      this.activeVideoElement_.playbackRate = rate;
+    }
+  }
+
+  /**
+   * mute active playback session
+   */
+  public mute(): void {
+    if (this.activeVideoElement_ && !this.activeVideoElement_.muted) {
+      this.activeVideoElement_.muted = true;
+      this.eventEmitter_.emitEvent(new MutedStatusChangedEvent(true));
+    }
+  }
+
+  /**
+   * unmute active playback session
+   */
+  public unmute(): void {
+    if (this.activeVideoElement_ && this.activeVideoElement_.muted) {
+      this.activeVideoElement_.muted = false;
+      this.eventEmitter_.emitEvent(new MutedStatusChangedEvent(false));
+    }
+  }
+
+  /**
+   * muted status getter
+   */
+  public getIsMuted(): boolean {
+    return this.activeVideoElement_?.muted ?? false;
+  }
+
+  /**
+   * volume level getter
+   */
+  public getVolumeLevel(): number {
+    return this.activeVideoElement_?.volume ?? 0;
+  }
+
+  /**
+   * update volume level for the active playback session
+   * @param volumeLevel - volume level
+   */
+  public setVolumeLevel(volumeLevel: number): void {
+    if (!this.activeVideoElement_) {
+      return;
+    }
+
+    let level: number;
+
+    if (volumeLevel > 1) {
+      level = 1;
+      this.logger_.warn(`Volume level must be in the [0, 1] range. Received: ${volumeLevel}. Value is clamped to 1.`);
+    } else if (volumeLevel < 0) {
+      level = 0;
+      this.logger_.warn(`Volume level must be in the [0, 1] range. Received: ${volumeLevel}. Value is clamped to 0.`);
+    } else {
+      level = volumeLevel;
+    }
+
+    this.activeVideoElement_.volume = level;
+  }
+
+  /**
+   * current playback session duration getter
+   */
+  public getDuration(): number {
+    // TODO: should we use (active seekable.end - active seekable.start)?
+    return this.activeVideoElement_?.duration ?? 0;
+  }
+
+  /**
+   * Current playback session state getter
+   */
+  public getPlaybackState(): PlaybackState {
+    return this.currentPlaybackState_;
+  }
+
+  /**
+   * get snapshot of the seekable ranges from the active pipeline
+   */
+  public getSeekableRanges(): Array<IPlayerTimeRange> {
+    if (this.activeVideoElement_) {
+      return PlayerTimeRange.fromTimeRanges(this.activeVideoElement_.seekable);
+    }
+
+    return [];
+  }
+
+  /**
+   * get snapshot of the current active seekable range:
+   * current time is in range inclusively
+   */
+  public getActiveSeekableRange(): IPlayerTimeRange | null {
+    return this.getSeekableRanges().find((range) => range.isInRangeInclusive(this.getCurrentTime())) ?? null;
+  }
+
+  /**
+   * get snapshot of the buffered ranges from the active pipeline
+   */
+  public getBufferedRanges(): Array<IPlayerTimeRange> {
+    if (this.activeVideoElement_) {
+      return PlayerTimeRange.fromTimeRanges(this.activeVideoElement_.buffered);
+    }
+
+    return [];
+  }
+
+  /**
+   * get snapshot of the current active buffered range:
+   * current time is in range inclusively
+   */
+  public getActiveBufferedRange(): IPlayerTimeRange | null {
+    return this.getBufferedRanges().find((range) => range.isInRangeInclusive(this.getCurrentTime())) ?? null;
+  }
+
+  /**
+   * Active playback session stats getter
+   */
+  public getPlaybackStats(): IPlaybackStats {
+    let droppedVideoFrames = 0;
+    let totalVideoFrames = 0;
+
+    if (this.activeVideoElement_) {
+      const playbackQuality = this.activeVideoElement_.getVideoPlaybackQuality();
+      droppedVideoFrames = playbackQuality.droppedVideoFrames;
+      totalVideoFrames = playbackQuality.totalVideoFrames;
+    }
+
+    // TODO: possibly collect other data from loader/pipeline/network
+    return {
+      droppedVideoFrames,
+      totalVideoFrames,
+    };
+  }
+
+  /**
+   * current playback session audio tracks getter
+   */
+  public getAudioTracks(): Array<IPlayerAudioTrack> {
+    // TODO: audio tracks
+    return [];
+  }
+
+  /**
+   * Active audio track getter
+   */
+  public getActiveAudioTrack(): IPlayerAudioTrack | null {
+    return this.getAudioTracks().find((track) => track.isActive) ?? null;
+  }
+
+  /**
+   * select audio track by id
+   * @param id - track id
+   */
+  public selectAudioTrack(id: string): boolean {
+    // TODO: select audio tracks
+    return Boolean(id);
+  }
+
+  /**
+   * current playback session thumbnail tracks getter
+   */
+  public getThumbnailTracks(): Array<IPlayerThumbnailTrack> {
+    // TODO: thumbnails tracks
+    return [];
+  }
+
+  /**
+   * Active thumbnail track getter
+   */
+  public getActiveThumbnailTrack(): IPlayerThumbnailTrack | null {
+    return this.getThumbnailTracks().find((track) => track.isActive) ?? null;
+  }
+
+  /**
+   * select thumbnails track by id
+   * @param id - track id
+   */
+  public selectThumbnailTrack(id: string): boolean {
+    // TODO: select thumbnails tracks
+    return Boolean(id);
+  }
+
+  /**
+   * add remote vtt thumbnail track (only VOD), should return false for live
+   * @param options - options
+   */
+  public addRemoteVttThumbnailTrack(options: IRemoteVttThumbnailTrackOptions): boolean {
+    // TODO: add remote vtt thumbnail track
+    return Boolean(options);
+  }
+
+  /**
+   * remove remote vtt thumbnail trac (only VOD), should return false for live
+   * @param id - track id
+   */
+  public removeRemoteVttThumbnailTrack(id: string): boolean {
+    // TODO: remove remote vtt thumbnail track
+    return Boolean(id);
+  }
+
+  /**
+   * current playback session quality levels getter
+   */
+  public getQualityLevels(): Array<IQualityLevel> {
+    // TODO: quality levels
+    return [];
+  }
+
+  /**
+   * Active quality level getter
+   */
+  public getActiveQualityLevel(): IQualityLevel | null {
+    return this.getQualityLevels().find((qualityLevel) => qualityLevel.isActive) ?? null;
+  }
+
+  /**
+   * select specific quality level, this will disable ABR
+   * @param id - quality level id
+   */
+  public selectQualityLevel(id: string): boolean {
+    // TODO: select quality level
+    return Boolean(id);
+  }
+
+  /**
+   * enable ABR
+   */
+  public selectAutoQualityLevel(): boolean {
+    // TODO: select auto quality level
+    return false;
+  }
 }
