+++ packages/playback/src/lib/types/version-info.declarations.ts
@@ -1,0 +1,5 @@
+export interface VersionInfo {
+  version: string;
+  versionHash: string;
+  isExperimental: boolean;
+}
