+++ packages/playback/test/player.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it } from 'vitest';
-import { Player } from '../src/lib/player';
+import { Player } from '../src/lib/player/main-thread/player-main';
 import {
   ConfigurationChangedEvent,
   LoggerLevelChangedEvent,
@@ -21,9 +21,7 @@ describe('Player spec', () => {
 
   describe('getLoggerLevel', () => {
     it('should return current logger level', () => {
-      // when(mockedLogger.getLoggerLevel()).thenReturn(Player.LoggerLevel.Warn);
       expect(player.getLoggerLevel()).toBe(Player.LoggerLevel.Debug);
-      // verify(mockedLogger.getLoggerLevel()).once();
     });
   });
 
@@ -36,7 +34,7 @@ describe('Player spec', () => {
 
       const actualEvents: Array<LoggerLevelChangedEvent> = [];
 
-      player.addEventListener(Player.Event.LoggerLevelChanged, (event) => {
+      player.addEventListener(Player.EventType.LoggerLevelChanged, (event) => {
         actualEvents.push(event);
       });
 
@@ -66,7 +64,7 @@ describe('Player spec', () => {
       const expectedEvents: Array<ConfigurationChangedEvent> = [];
       const actualEvents: Array<ConfigurationChangedEvent> = [];
 
-      player.addEventListener(Player.Event.ConfigurationChanged, (event) => {
+      player.addEventListener(Player.EventType.ConfigurationChanged, (event) => {
         actualEvents.push(event);
       });
 
@@ -94,7 +92,7 @@ describe('Player spec', () => {
       const expectedEvents: Array<ConfigurationChangedEvent> = [];
       const actualEvents: Array<ConfigurationChangedEvent> = [];
 
-      player.addEventListener(Player.Event.ConfigurationChanged, (event) => {
+      player.addEventListener(Player.EventType.ConfigurationChanged, (event) => {
         actualEvents.push(event);
       });
 
@@ -132,11 +130,11 @@ describe('Player spec', () => {
       ];
       const actualEvents: Array<PlayerEvent> = [];
 
-      player.addEventListener(Player.Event.LoggerLevelChanged, (event) => {
+      player.addEventListener(Player.EventType.LoggerLevelChanged, (event) => {
         actualEvents.push(event);
       });
 
-      player.addEventListener(Player.Event.VolumeChanged, (event) => {
+      player.addEventListener(Player.EventType.VolumeChanged, (event) => {
         actualEvents.push(event);
       });
 
@@ -154,7 +152,7 @@ describe('Player spec', () => {
       const expectedEvents = [new LoggerLevelChangedEvent(Player.LoggerLevel.Warn)];
       const actualEvents: Array<PlayerEvent> = [];
 
-      player.once(Player.Event.LoggerLevelChanged, (event) => {
+      player.once(Player.EventType.LoggerLevelChanged, (event) => {
         actualEvents.push(event);
       });
 
@@ -176,17 +174,17 @@ describe('Player spec', () => {
         actualEvents.push(event);
       };
 
-      player.addEventListener(Player.Event.LoggerLevelChanged, listener);
-      player.addEventListener(Player.Event.VolumeChanged, listener);
+      player.addEventListener(Player.EventType.LoggerLevelChanged, listener);
+      player.addEventListener(Player.EventType.VolumeChanged, listener);
 
       serviceLocator.eventEmitter.emitEvent(new LoggerLevelChangedEvent(Player.LoggerLevel.Warn));
 
-      player.removeEventListener(Player.Event.LoggerLevelChanged, listener);
+      player.removeEventListener(Player.EventType.LoggerLevelChanged, listener);
 
       serviceLocator.eventEmitter.emitEvent(new LoggerLevelChangedEvent(Player.LoggerLevel.Info));
       serviceLocator.eventEmitter.emitEvent(new VolumeChangedEvent(0.5));
 
-      player.removeEventListener(Player.Event.VolumeChanged, listener);
+      player.removeEventListener(Player.EventType.VolumeChanged, listener);
 
       serviceLocator.eventEmitter.emitEvent(new VolumeChangedEvent(0.6));
 
@@ -217,13 +215,13 @@ describe('Player spec', () => {
         actualEvents.push(event);
       };
 
-      player.addEventListener(Player.Event.LoggerLevelChanged, listener1);
-      player.addEventListener(Player.Event.LoggerLevelChanged, listener2);
-      player.addEventListener(Player.Event.LoggerLevelChanged, listener3);
+      player.addEventListener(Player.EventType.LoggerLevelChanged, listener1);
+      player.addEventListener(Player.EventType.LoggerLevelChanged, listener2);
+      player.addEventListener(Player.EventType.LoggerLevelChanged, listener3);
 
       serviceLocator.eventEmitter.emitEvent(new LoggerLevelChangedEvent(Player.LoggerLevel.Info));
 
-      player.removeAllEventListenersForType(Player.Event.LoggerLevelChanged);
+      player.removeAllEventListenersForType(Player.EventType.LoggerLevelChanged);
 
       serviceLocator.eventEmitter.emitEvent(new LoggerLevelChangedEvent(Player.LoggerLevel.Info));
       serviceLocator.eventEmitter.emitEvent(new LoggerLevelChangedEvent(Player.LoggerLevel.Info));
@@ -242,8 +240,8 @@ describe('Player spec', () => {
         actualEvents.push(event);
       };
 
-      player.addEventListener(Player.Event.LoggerLevelChanged, listener);
-      player.addEventListener(Player.Event.VolumeChanged, listener);
+      player.addEventListener(Player.EventType.LoggerLevelChanged, listener);
+      player.addEventListener(Player.EventType.VolumeChanged, listener);
 
       serviceLocator.eventEmitter.emitEvent(new LoggerLevelChangedEvent(Player.LoggerLevel.Warn));
       serviceLocator.eventEmitter.emitEvent(new VolumeChangedEvent(0.5));
