+++ packages/playback/test/utils/interceptors-storage.test.ts
@@ -1,9 +1,10 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 import { InterceptorsStorage } from '../../src/lib/utils/interceptors-storage';
 import { InterceptorType } from '../../src/lib/consts/interceptor-type';
+import type { InterceptorTypeToInterceptorPayloadMap } from '../../src';
 
 describe('InterceptorsStorage', () => {
-  let storage: InterceptorsStorage;
+  let storage: InterceptorsStorage<InterceptorTypeToInterceptorPayloadMap>;
 
   beforeEach(() => {
     storage = new InterceptorsStorage();
